# docs/api/blueiq-api-gateway-openapi.yaml
openapi: 3.0.3
info:
  title: BlueIQ API Gateway
  version: 1.0.0
  contact:
    name: API Support
    url: http://www.example.com/support
    email: support@example.com
  description: |
    This document describes the API for the BlueIQ API Gateway.
    It serves as the single entry point for all client interactions with the BlueIQ microservices.
    This specification is designed to be a reference for consumers and a template for similar gateway patterns.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://dev.blueiq.example.com/api/v1
    description: Development environment
  - url: https://staging.blueiq.example.com/api/v1
    description: Staging environment
  - url: https://blueiq.example.com/api/v1
    description: Production environment
security:
  - BearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Log in a user
      operationId: loginUser
      description: Authenticates a user with their credentials and returns JWT access and refresh tokens.
      # No global security for login, it's public
      security: [] 
      requestBody:
        description: User credentials for login
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Login successful, returns access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '400':
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh an access token
      operationId: refreshToken
      description: Obtains a new JWT access token using a valid refresh token.
      # No global security for refresh, it's public
      security: [] 
      requestBody:
        description: Refresh token
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Log out a user
      operationId: logoutUser
      description: Invalidates the user's current session. The specifics of invalidation (e.g., token blocklisting) depend on server implementation.
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Invalid input provided.
        details:
          type: string
          nullable: true
          example: The 'username' field is required.
    UserLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: adminuser
          description: The user's username.
        password:
          type: string
          example: 'P@$$wOrd123!'
          description: The user's password.
    LoginSuccessResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
          description: JWT access token.
        refreshToken:
          type: string
          example: def50200f07711e9b210d663bd873d93
          description: Token used to refresh the access token.
        expiresIn:
          type: integer
          example: 3600
          description: Access token validity period in seconds.
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: def50200f07711e9b210d663bd873d93
          description: The refresh token.
...
  responses:
    UnauthorizedError:
      description: Unauthorized - Authentication token is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing token
              value:
                statusCode: 401
                message: "Authentication token is missing."
            invalidToken:
              summary: Invalid token
              value:
                statusCode: 401
                message: "Authentication token is invalid or expired."
...
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            genericError:
              summary: Generic server error
              value:
                statusCode: 500
                message: "An unexpected error occurred. Please try again later or contact support."
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token for authentication. The token should be included in the
        Authorization header as 'Bearer {token}'.
        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`